test: simplify test file hierarchy
